Fix warnings in raylib (#3793)

+++ src/rcore.c
@@ -1406,7 +1406,7 @@ void SetShaderValueTexture(Shader shader, int locIndex, Texture2D texture)
 // Get a ray trace from mouse position
 Ray GetMouseRay(Vector2 mousePosition, Camera camera)
 {
-    return GetViewRay(mousePosition, camera, GetScreenWidth(), GetScreenHeight());
+    return GetViewRay(mousePosition, camera, (float)GetScreenWidth(), (float)GetScreenHeight());
 }
 
 // Get a ray trace from the mouse position within a specific section of the screen
